Added http response checking
Found that no exceptions are raised if the http.get() function returns a 401 or other http status codes that indicate a failed request. This would have us store the 401 page or other error pages set by the server into our database, which remains there until either we clear the cache or the tiles time out.

I added in some code to check the response from the http.get() function to assure the request statusCode is 200 which indicates a successful response from the server. This was done in both the loadTiles() and _loadAsync() functions so that the same behavior occurs in both cases, Now if the server responds with something other than a valid tile, we don't store this in the database.