Added HTTP response checking (#5)
Found that no exceptions are raised if the http.get() function returns a 401 or other http status codes that indicate a failed request. This would have us store the 401 page or other error pages set by the server into our database, which remains there until either we clear the cache or the tiles time out.